@@ -1,11 +1,11 @@
 package com.example.yangzhe.bottomtablayout.adapter;
 
 import android.content.Context;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
-import android.support.v4.view.LayoutInflaterCompat;
-import android.support.v4.view.ViewPager;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
+import androidx.core.view.LayoutInflaterCompat;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.ImageView;
