@@ -1,6 +1,6 @@
 package com.example.yangzhe.coroutine
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import com.example.yangzhe.learnactivity.R
