@@ -1,7 +1,7 @@
 package com.example.yangzhe.learnipc
 
 import android.content.Intent
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import com.example.yangzhe.learnactivity.R
